Fixed a large number of WIN64 type conversion warnings.
Changed FS_Write to do multiple write calls if the size exceeds 1GB, this is to accommodate WIN64 where write() takes a UINT rather than size_t.  We have no use of such large writes at this time, so this is a mostly theoretical feature.

git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@12081 d7cf8633-e32d-0410-b094-e92efae38249
::stable-branch::merge=d090340b054029faddb432e2c66085e6e91abcf7

+++ common.c
@@ -2287,7 +2287,7 @@ size_t base64_encode(unsigned char *buf, size_t buflen, size_t outbuflen)
 	for(i = blocks; i > 0; )
 	{
 		--i;
-		base64_3to4(buf + 3*i, buf + 4*i, buflen - 3*i);
+		base64_3to4(buf + 3*i, buf + 4*i, (int)(buflen - 3*i));
 	}
 	return blocks * 4;
 }
